dev: add madara genesis and move to folder
